Struct redbpf_probes::bindings::bus_type

#[repr(C)]
pub struct bus_type {
    pub name: *const c_char,
    pub dev_name: *const c_char,
    pub dev_root: *mut device,
    pub bus_groups: *mut *const attribute_group,
    pub dev_groups: *mut *const attribute_group,
    pub drv_groups: *mut *const attribute_group,
    pub match_: Option<unsafe extern "C" fn(dev: *mut device, drv: *mut device_driver) -> c_int>,
    pub uevent: Option<unsafe extern "C" fn(dev: *mut device, env: *mut kobj_uevent_env) -> c_int>,
    pub probe: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub remove: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub shutdown: Option<unsafe extern "C" fn(dev: *mut device)>,
    pub online: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub offline: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub suspend: Option<unsafe extern "C" fn(dev: *mut device, state: pm_message_t) -> c_int>,
    pub resume: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub num_vf: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub pm: *const dev_pm_ops,
    pub iommu_ops: *const iommu_ops,
    pub p: *mut subsys_private,
    pub lock_key: lock_class_key,
    pub force_dma: bool_,
}

struct bus_type - The bus type of the device

@name: The name of the bus. @dev_name: Used for subsystems to enumerate devices like ("foo%u", dev->id). @dev_root: Default device to use as the parent. @bus_groups: Default attributes of the bus. @dev_groups: Default attributes of the devices on the bus. @drv_groups: Default attributes of the device drivers on the bus. @match: Called, perhaps multiple times, whenever a new device or driver is added for this bus. It should return a positive value if the given device can be handled by the given driver and zero otherwise. It may also return error code if determining that the driver supports the device is not possible. In case of -EPROBE_DEFER it will queue the device for deferred probing. @uevent: Called when a device is added, removed, or a few other things that generate uevents to add the environment variables. @probe: Called when a new device or driver add to this bus, and callback the specific driver's probe to initial the matched device. @remove: Called when a device removed from this bus. @shutdown: Called at shut-down time to quiesce the device.

@online: Called to put the device back online (after offlining it). @offline: Called to put the device offline for hot-removal. May fail.

@suspend: Called when a device on this bus wants to go to sleep mode. @resume: Called to bring a device on this bus out of sleep mode. @num_vf: Called to find out how many virtual functions a device on this bus supports. @pm: Power management operations of this bus, callback the specific device driver's pm-ops. @iommu_ops: IOMMU specific operations for this bus, used to attach IOMMU driver implementations to a bus and allow the driver to do bus-specific setup @p: The private data of the driver core, only the driver core can touch this. @lock_key: Lock class key for use by the lock validator @force_dma: Assume devices on this bus should be set up by dma_configure() even if DMA capability is not explicitly described by firmware.

A bus is a channel between the processor and one or more devices. For the purposes of the device model, all devices are connected via a bus, even if it is an internal, virtual, "platform" bus. Buses can plug into each other. A USB controller is usually a PCI device, for example. The device model represents the actual connections between buses and the devices they control. A bus is represented by the bus_type structure. It contains the name, the default attributes, the bus' methods, PM operations, and the driver core's private data.

Fields
name: *const c_chardev_name: *const c_chardev_root: *mut devicebus_groups: *mut *const attribute_groupdev_groups: *mut *const attribute_groupdrv_groups: *mut *const attribute_groupmatch_: Option<unsafe extern "C" fn(dev: *mut device, drv: *mut device_driver) -> c_int>uevent: Option<unsafe extern "C" fn(dev: *mut device, env: *mut kobj_uevent_env) -> c_int>probe: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>remove: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>shutdown: Option<unsafe extern "C" fn(dev: *mut device)>online: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>offline: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>suspend: Option<unsafe extern "C" fn(dev: *mut device, state: pm_message_t) -> c_int>resume: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>num_vf: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>pm: *const dev_pm_opsiommu_ops: *const iommu_opsp: *mut subsys_privatelock_key: lock_class_keyforce_dma: bool_