Struct redbpf_probes::bindings::class

#[repr(C)]
pub struct class {
    pub name: *const c_char,
    pub owner: *mut module,
    pub class_groups: *mut *const attribute_group,
    pub dev_groups: *mut *const attribute_group,
    pub dev_kobj: *mut kobject,
    pub dev_uevent: Option<unsafe extern "C" fn(dev: *mut device, env: *mut kobj_uevent_env) -> c_int>,
    pub devnode: Option<unsafe extern "C" fn(dev: *mut device, mode: *mut umode_t) -> *mut c_char>,
    pub class_release: Option<unsafe extern "C" fn(class: *mut class)>,
    pub dev_release: Option<unsafe extern "C" fn(dev: *mut device)>,
    pub shutdown_pre: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub ns_type: *const kobj_ns_type_operations,
    pub namespace: Option<unsafe extern "C" fn(dev: *mut device) -> *const c_void>,
    pub pm: *const dev_pm_ops,
    pub p: *mut subsys_private,
}

struct class - device classes @name: Name of the class. @owner: The module owner. @class_groups: Default attributes of this class. @dev_groups: Default attributes of the devices that belong to the class. @dev_kobj: The kobject that represents this class and links it into the hierarchy. @dev_uevent: Called when a device is added, removed from this class, or a few other things that generate uevents to add the environment variables. @devnode: Callback to provide the devtmpfs. @class_release: Called to release this class. @dev_release: Called to release the device. @shutdown_pre: Called at shut-down time before driver shutdown. @ns_type: Callbacks so sysfs can detemine namespaces. @namespace: Namespace of the device belongs to this class. @pm: The default device power management operations of this class. @p: The private data of the driver core, no one other than the driver core can touch this.

A class is a higher-level view of a device that abstracts out low-level implementation details. Drivers may see a SCSI disk or an ATA disk, but, at the class level, they are all simply disks. Classes allow user space to work with devices based on what they do, rather than how they are connected or how they work.

Fields
name: *const c_charowner: *mut moduleclass_groups: *mut *const attribute_groupdev_groups: *mut *const attribute_groupdev_kobj: *mut kobjectdev_uevent: Option<unsafe extern "C" fn(dev: *mut device, env: *mut kobj_uevent_env) -> c_int>devnode: Option<unsafe extern "C" fn(dev: *mut device, mode: *mut umode_t) -> *mut c_char>class_release: Option<unsafe extern "C" fn(class: *mut class)>dev_release: Option<unsafe extern "C" fn(dev: *mut device)>shutdown_pre: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>ns_type: *const kobj_ns_type_operationsnamespace: Option<unsafe extern "C" fn(dev: *mut device) -> *const c_void>pm: *const dev_pm_opsp: *mut subsys_private