Struct redbpf_probes::bindings::device

#[repr(C)]
pub struct device {
    pub parent: *mut device,
    pub p: *mut device_private,
    pub kobj: kobject,
    pub init_name: *const c_char,
    pub type_: *const device_type,
    pub mutex: mutex,
    pub bus: *mut bus_type,
    pub driver: *mut device_driver,
    pub platform_data: *mut c_void,
    pub driver_data: *mut c_void,
    pub links: dev_links_info,
    pub power: dev_pm_info,
    pub pm_domain: *mut dev_pm_domain,
    pub msi_domain: *mut irq_domain,
    pub pins: *mut dev_pin_info,
    pub msi_list: list_head,
    pub numa_node: c_int,
    pub dma_ops: *const dma_map_ops,
    pub dma_mask: *mut u64,
    pub coherent_dma_mask: u64,
    pub dma_pfn_offset: c_ulong,
    pub dma_parms: *mut device_dma_parameters,
    pub dma_pools: list_head,
    pub dma_mem: *mut dma_coherent_mem,
    pub archdata: dev_archdata,
    pub of_node: *mut device_node,
    pub fwnode: *mut fwnode_handle,
    pub devt: dev_t,
    pub id: u32,
    pub devres_lock: spinlock_t,
    pub devres_head: list_head,
    pub knode_class: klist_node,
    pub class: *mut class,
    pub groups: *mut *const attribute_group,
    pub release: Option<unsafe extern "C" fn(dev: *mut device)>,
    pub iommu_group: *mut iommu_group,
    pub iommu_fwspec: *mut iommu_fwspec,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>,
    pub __bindgen_padding_0: [u8; 7],
}

struct device - The basic device structure @parent: The device's "parent" device, the device to which it is attached. In most cases, a parent device is some sort of bus or host controller. If parent is NULL, the device, is a top-level device, which is not usually what you want. @p: Holds the private data of the driver core portions of the device. See the comment of the struct device_private for detail. @kobj: A top-level, abstract class from which other classes are derived. @init_name: Initial name of the device. @type: The type of device. This identifies the device type and carries type-specific information. @mutex: Mutex to synchronize calls to its driver. @bus: Type of bus device is on. @driver: Which driver has allocated this @platform_data: Platform data specific to the device. Example: For devices on custom boards, as typical of embedded and SOC based hardware, Linux often uses platform_data to point to board-specific structures describing devices and how they are wired. That can include what ports are available, chip variants, which GPIO pins act in what additional roles, and so on. This shrinks the "Board Support Packages" (BSPs) and minimizes board-specific #ifdefs in drivers. @driver_data: Private pointer for driver specific info. @links: Links to suppliers and consumers of this device. @power: For device power management. See Documentation/driver-api/pm/devices.rst for details. @pm_domain: Provide callbacks that are executed during system suspend, hibernation, system resume and during runtime PM transitions along with subsystem-level and driver-level callbacks. @pins: For device pin management. See Documentation/driver-api/pinctl.rst for details. @msi_list: Hosts MSI descriptors @msi_domain: The generic MSI domain this device is using. @numa_node: NUMA node this device is close to. @dma_ops: DMA mapping operations for this device. @dma_mask: Dma mask (if dma'ble device). @coherent_dma_mask: Like dma_mask, but for alloc_coherent mapping as not all hardware supports 64-bit addresses for consistent allocations such descriptors. @dma_pfn_offset: offset of DMA memory range relatively of RAM @dma_parms: A low level driver may set these to teach IOMMU code about segment limitations. @dma_pools: Dma pools (if dma'ble device). @dma_mem: Internal for coherent mem override. @cma_area: Contiguous memory area for dma allocations @archdata: For arch-specific additions. @of_node: Associated device tree node. @fwnode: Associated device node supplied by platform firmware. @devt: For creating the sysfs "dev". @id: device instance @devres_lock: Spinlock to protect the resource of the device. @devres_head: The resources list of the device. @knode_class: The node used to add the device to the class list. @class: The class of the device. @groups: Optional attribute groups. @release: Callback to free the device after all references have gone away. This should be set by the allocator of the device (i.e. the bus driver that discovered the device). @iommu_group: IOMMU group the device belongs to. @iommu_fwspec: IOMMU-specific properties supplied by firmware.

@offline_disabled: If set, the device is permanently online. @offline: Set after successful invocation of bus type's .offline(). @of_node_reused: Set if the device-tree node is shared with an ancestor device.

At the lowest level, every device in a Linux system is represented by an instance of struct device. The device structure contains the information that the device model core needs to model the system. Most subsystems, however, track additional information about the devices they host. As a result, it is rare for devices to be represented by bare device structures; instead, that structure, like kobject structures, is usually embedded within a higher-level representation of the device.

Fields
parent: *mut devicep: *mut device_privatekobj: kobjectinit_name: *const c_chartype_: *const device_typemutex: mutexbus: *mut bus_typedriver: *mut device_driverplatform_data: *mut c_voiddriver_data: *mut c_voidlinks: dev_links_infopower: dev_pm_infopm_domain: *mut dev_pm_domainmsi_domain: *mut irq_domainpins: *mut dev_pin_infomsi_list: list_headnuma_node: c_intdma_ops: *const dma_map_opsdma_mask: *mut u64coherent_dma_mask: u64dma_pfn_offset: c_ulongdma_parms: *mut device_dma_parametersdma_pools: list_headdma_mem: *mut dma_coherent_memarchdata: dev_archdataof_node: *mut device_nodefwnode: *mut fwnode_handledevt: dev_tid: u32devres_lock: spinlock_tdevres_head: list_headknode_class: klist_nodeclass: *mut classgroups: *mut *const attribute_grouprelease: Option<unsafe extern "C" fn(dev: *mut device)>iommu_group: *mut iommu_groupiommu_fwspec: *mut iommu_fwspec_bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>__bindgen_padding_0: [u8; 7]
Methods

impl device

pub fn offline_disabled(&self) -> bool_

pub fn set_offline_disabled(&mut self, val: bool_)

pub fn offline(&self) -> bool_

pub fn set_offline(&mut self, val: bool_)

pub fn of_node_reused(&self) -> bool_

pub fn set_of_node_reused(&mut self, val: bool_)

pub fn new_bitfield_1(
offline_disabled: bool_,
offline: bool_,
of_node_reused: bool_
) -> __BindgenBitfieldUnit<[u8; 1], u8>