Struct redbpf_probes::bindings::device_driver

#[repr(C)]
pub struct device_driver {
    pub name: *const c_char,
    pub bus: *mut bus_type,
    pub owner: *mut module,
    pub mod_name: *const c_char,
    pub suppress_bind_attrs: bool_,
    pub probe_type: probe_type,
    pub of_match_table: *mut of_device_id,
    pub acpi_match_table: *mut acpi_device_id,
    pub probe: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub remove: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub shutdown: Option<unsafe extern "C" fn(dev: *mut device)>,
    pub suspend: Option<unsafe extern "C" fn(dev: *mut device, state: pm_message_t) -> c_int>,
    pub resume: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub groups: *mut *const attribute_group,
    pub pm: *const dev_pm_ops,
    pub coredump: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>,
    pub p: *mut driver_private,
}

struct device_driver - The basic device driver structure @name: Name of the device driver. @bus: The bus which the device of this driver belongs to. @owner: The module owner. @mod_name: Used for built-in modules. @suppress_bind_attrs: Disables bind/unbind via sysfs. @probe_type: Type of the probe (synchronous or asynchronous) to use. @of_match_table: The open firmware table. @acpi_match_table: The ACPI match table. @probe: Called to query the existence of a specific device, whether this driver can work with it, and bind the driver to a specific device. @remove: Called when the device is removed from the system to unbind a device from this driver. @shutdown: Called at shut-down time to quiesce the device. @suspend: Called to put the device to sleep mode. Usually to a low power state. @resume: Called to bring a device from sleep mode. @groups: Default attributes that get created by the driver core automatically. @pm: Power management operations of the device which matched this driver. @p: Driver core's private data, no one other than the driver core can touch this.

The device driver-model tracks all of the drivers known to the system. The main reason for this tracking is to enable the driver core to match up drivers with new devices. Once drivers are known objects within the system, however, a number of other things become possible. Device drivers can export information and configuration variables that are independent of any specific device.

Fields
name: *const c_charbus: *mut bus_typeowner: *mut modulemod_name: *const c_charsuppress_bind_attrs: bool_probe_type: probe_typeof_match_table: *mut of_device_idacpi_match_table: *mut acpi_device_idprobe: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>remove: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>shutdown: Option<unsafe extern "C" fn(dev: *mut device)>suspend: Option<unsafe extern "C" fn(dev: *mut device, state: pm_message_t) -> c_int>resume: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>groups: *mut *const attribute_grouppm: *const dev_pm_opscoredump: Option<unsafe extern "C" fn(dev: *mut device) -> c_int>p: *mut driver_private