Struct redbpf_probes::bindings::ethtool_link_settings

#[repr(C)]
pub struct ethtool_link_settings {
    pub cmd: __u32,
    pub speed: __u32,
    pub duplex: __u8,
    pub port: __u8,
    pub phy_address: __u8,
    pub autoneg: __u8,
    pub mdio_support: __u8,
    pub eth_tp_mdix: __u8,
    pub eth_tp_mdix_ctrl: __u8,
    pub link_mode_masks_nwords: __s8,
    pub transceiver: __u8,
    pub reserved1: [__u8; 3],
    pub reserved: [__u32; 7],
    pub link_mode_masks: __IncompleteArrayField<__u32>,
}

struct ethtool_link_settings - link control and status

IMPORTANT, Backward compatibility notice: When implementing new user-space tools, please first try %ETHTOOL_GLINKSETTINGS, and if it succeeds use %ETHTOOL_SLINKSETTINGS to change link settings; do not use %ETHTOOL_SSET if %ETHTOOL_GLINKSETTINGS succeeded: stick to %ETHTOOL_GLINKSETTINGS/%SLINKSETTINGS in that case. Conversely, if %ETHTOOL_GLINKSETTINGS fails, use %ETHTOOL_GSET to query and %ETHTOOL_SSET to change link settings; do not use %ETHTOOL_SLINKSETTINGS if %ETHTOOL_GLINKSETTINGS failed: stick to %ETHTOOL_GSET/%ETHTOOL_SSET in that case.

@cmd: Command number = %ETHTOOL_GLINKSETTINGS or %ETHTOOL_SLINKSETTINGS @speed: Link speed (Mbps) @duplex: Duplex mode; one of %DUPLEX_* @port: Physical connector type; one of %PORT_* @phy_address: MDIO address of PHY (transceiver); 0 or 255 if not applicable. For clause 45 PHYs this is the PRTAD. @autoneg: Enable/disable autonegotiation and auto-detection; either %AUTONEG_DISABLE or %AUTONEG_ENABLE @mdio_support: Bitmask of %ETH_MDIO_SUPPORTS_* flags for the MDIO protocols supported by the interface; 0 if unknown. Read-only. @eth_tp_mdix: Ethernet twisted-pair MDI(-X) status; one of %ETH_TP_MDI_. If the status is unknown or not applicable, the value will be %ETH_TP_MDI_INVALID. Read-only. @eth_tp_mdix_ctrl: Ethernet twisted pair MDI(-X) control; one of %ETH_TP_MDI_. If MDI(-X) control is not implemented, reads yield %ETH_TP_MDI_INVALID and writes may be ignored or rejected. When written successfully, the link should be renegotiated if necessary. @link_mode_masks_nwords: Number of 32-bit words for each of the supported, advertising, lp_advertising link mode bitmaps. For %ETHTOOL_GLINKSETTINGS: on entry, number of words passed by user (>= 0); on return, if handshake in progress, negative if request size unsupported by kernel: absolute value indicates kernel expected size and all the other fields but cmd are 0; otherwise (handshake completed), strictly positive to indicate size used by kernel and cmd field stays %ETHTOOL_GLINKSETTINGS, all other fields populated by driver. For %ETHTOOL_SLINKSETTINGS: must be valid on entry, ie. a positive value returned previously by %ETHTOOL_GLINKSETTINGS, otherwise refused. For drivers: ignore this field (use kernel's __ETHTOOL_LINK_MODE_MASK_NBITS instead), any change to it will be overwritten by kernel. @supported: Bitmap with each bit meaning given by %ethtool_link_mode_bit_indices for the link modes, physical connectors and other link features for which the interface supports autonegotiation or auto-detection. Read-only. @advertising: Bitmap with each bit meaning given by %ethtool_link_mode_bit_indices for the link modes, physical connectors and other link features that are advertised through autonegotiation or enabled for auto-detection. @lp_advertising: Bitmap with each bit meaning given by %ethtool_link_mode_bit_indices for the link modes, and other link features that the link partner advertised through autonegotiation; 0 if unknown or not applicable. Read-only. @transceiver: Used to distinguish different possible PHY types, reported consistently by PHYLIB. Read-only.

If autonegotiation is disabled, the speed and @duplex represent the fixed link mode and are writable if the driver supports multiple link modes. If it is enabled then they are read-only; if the link is up they represent the negotiated link mode; if the link is down, the speed is 0, %SPEED_UNKNOWN or the highest enabled speed and @duplex is %DUPLEX_UNKNOWN or the best enabled duplex mode.

Some hardware interfaces may have multiple PHYs and/or physical connectors fitted or do not allow the driver to detect which are fitted. For these interfaces @port and/or @phy_address may be writable, possibly dependent on @autoneg being %AUTONEG_DISABLE. Otherwise, attempts to write different values may be ignored or rejected.

Deprecated %ethtool_cmd fields transceiver, maxtxpkt and maxrxpkt are not available in %ethtool_link_settings. Until all drivers are converted to ignore them or to the new %ethtool_link_settings API, for both queries and changes, users should always try %ETHTOOL_GLINKSETTINGS first, and if it fails with -ENOTSUPP stick only to %ETHTOOL_GSET and %ETHTOOL_SSET consistently. If it succeeds, then users should stick to %ETHTOOL_GLINKSETTINGS and %ETHTOOL_SLINKSETTINGS (which would support drivers implementing either %ethtool_cmd or %ethtool_link_settings).

Users should assume that all fields not marked read-only are writable and subject to validation by the driver. They should use %ETHTOOL_GLINKSETTINGS to get the current values before making specific changes and then applying them with %ETHTOOL_SLINKSETTINGS.

Drivers that implement %get_link_ksettings and/or %set_link_ksettings should ignore the @cmd and @link_mode_masks_nwords fields (any change to them overwritten by kernel), and rely only on kernel's internal %__ETHTOOL_LINK_MODE_MASK_NBITS and %ethtool_link_mode_mask_t. Drivers that implement %set_link_ksettings() should validate all fields other than @cmd and @link_mode_masks_nwords that are not described as read-only or deprecated, and must ignore all fields described as read-only.

Fields
cmd: __u32speed: __u32duplex: __u8port: __u8phy_address: __u8autoneg: __u8mdio_support: __u8eth_tp_mdix: __u8eth_tp_mdix_ctrl: __u8link_mode_masks_nwords: __s8transceiver: __u8reserved1: [__u8; 3]reserved: [__u32; 7]link_mode_masks: __IncompleteArrayField<__u32>