Struct redbpf_probes::bindings::ethtool_ops

#[repr(C)]pub struct ethtool_ops {
    pub get_drvinfo: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_drvinfo)>,
    pub get_regs_len: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>,
    pub get_regs: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_regs, arg3: *mut c_void)>,
    pub get_wol: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_wolinfo)>,
    pub set_wol: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_wolinfo) -> c_int>,
    pub get_msglevel: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>,
    pub set_msglevel: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32)>,
    pub nway_reset: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>,
    pub get_link: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>,
    pub get_eeprom_len: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>,
    pub get_eeprom: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eeprom, arg3: *mut _cargo_bpf_u8) -> c_int>,
    pub set_eeprom: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eeprom, arg3: *mut _cargo_bpf_u8) -> c_int>,
    pub get_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_coalesce) -> c_int>,
    pub set_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_coalesce) -> c_int>,
    pub get_ringparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_ringparam)>,
    pub set_ringparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_ringparam) -> c_int>,
    pub get_pauseparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_pauseparam)>,
    pub set_pauseparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_pauseparam) -> c_int>,
    pub self_test: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_test, arg3: *mut _cargo_bpf_u64)>,
    pub get_strings: Option<unsafe extern "C" fn(arg1: *mut net_device, stringset: _cargo_bpf_u32, arg2: *mut _cargo_bpf_u8)>,
    pub set_phys_id: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: ethtool_phys_id_state) -> c_int>,
    pub get_ethtool_stats: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_stats, arg3: *mut _cargo_bpf_u64)>,
    pub begin: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>,
    pub complete: Option<unsafe extern "C" fn(arg1: *mut net_device)>,
    pub get_priv_flags: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>,
    pub set_priv_flags: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32) -> c_int>,
    pub get_sset_count: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: c_int) -> c_int>,
    pub get_rxnfc: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_rxnfc, rule_locs: *mut _cargo_bpf_u32) -> c_int>,
    pub set_rxnfc: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_rxnfc) -> c_int>,
    pub flash_device: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_flash) -> c_int>,
    pub reset: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut _cargo_bpf_u32) -> c_int>,
    pub get_rxfh_key_size: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>,
    pub get_rxfh_indir_size: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>,
    pub get_rxfh: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *mut _cargo_bpf_u32, key: *mut _cargo_bpf_u8, hfunc: *mut _cargo_bpf_u8) -> c_int>,
    pub set_rxfh: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *const _cargo_bpf_u32, key: *const _cargo_bpf_u8, hfunc: _cargo_bpf_u8) -> c_int>,
    pub get_rxfh_context: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *mut _cargo_bpf_u32, key: *mut _cargo_bpf_u8, hfunc: *mut _cargo_bpf_u8, rss_context: _cargo_bpf_u32) -> c_int>,
    pub set_rxfh_context: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *const _cargo_bpf_u32, key: *const _cargo_bpf_u8, hfunc: _cargo_bpf_u8, rss_context: *mut _cargo_bpf_u32, delete: bool_) -> c_int>,
    pub get_channels: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_channels)>,
    pub set_channels: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_channels) -> c_int>,
    pub get_dump_flag: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_dump) -> c_int>,
    pub get_dump_data: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_dump, arg3: *mut c_void) -> c_int>,
    pub set_dump: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_dump) -> c_int>,
    pub get_ts_info: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_ts_info) -> c_int>,
    pub get_module_info: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_modinfo) -> c_int>,
    pub get_module_eeprom: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eeprom, arg3: *mut _cargo_bpf_u8) -> c_int>,
    pub get_eee: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eee) -> c_int>,
    pub set_eee: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eee) -> c_int>,
    pub get_tunable: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *const ethtool_tunable, arg3: *mut c_void) -> c_int>,
    pub set_tunable: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *const ethtool_tunable, arg3: *const c_void) -> c_int>,
    pub get_per_queue_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32, arg3: *mut ethtool_coalesce) -> c_int>,
    pub set_per_queue_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32, arg3: *mut ethtool_coalesce) -> c_int>,
    pub get_link_ksettings: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_link_ksettings) -> c_int>,
    pub set_link_ksettings: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *const ethtool_link_ksettings) -> c_int>,
    pub get_fecparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_fecparam) -> c_int>,
    pub set_fecparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_fecparam) -> c_int>,
    pub get_ethtool_phy_stats: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_stats, arg3: *mut _cargo_bpf_u64)>,
}

struct ethtool_ops - optional netdev operations @get_drvinfo: Report driver/device information. Should only set the @driver, @version, @fw_version and @bus_info fields. If not implemented, the @driver and @bus_info fields will be filled in according to the netdev's parent device. @get_regs_len: Get buffer length required for @get_regs @get_regs: Get device registers @get_wol: Report whether Wake-on-Lan is enabled @set_wol: Turn Wake-on-Lan on or off. Returns a negative error code or zero. @get_msglevel: Report driver message level. This should be the value of the @msg_enable field used by netif logging functions. @set_msglevel: Set driver message level @nway_reset: Restart autonegotiation. Returns a negative error code or zero. @get_link: Report whether physical link is up. Will only be called if the netdev is up. Should usually be set to ethtool_op_get_link(), which uses netif_carrier_ok(). @get_eeprom: Read data from the device EEPROM. Should fill in the magic field. Don't need to check len for zero or wraparound. Fill in the data argument with the eeprom values from offset to offset + len. Update len to the amount read. Returns an error or zero. @set_eeprom: Write data to the device EEPROM. Should validate the magic field. Don't need to check len for zero or wraparound. Update len to the amount written. Returns an error or zero. @get_coalesce: Get interrupt coalescing parameters. Returns a negative error code or zero. @set_coalesce: Set interrupt coalescing parameters. Returns a negative error code or zero. @get_ringparam: Report ring sizes @set_ringparam: Set ring sizes. Returns a negative error code or zero. @get_pauseparam: Report pause parameters @set_pauseparam: Set pause parameters. Returns a negative error code or zero. @self_test: Run specified self-tests @get_strings: Return a set of strings that describe the requested objects @set_phys_id: Identify the physical devices, e.g. by flashing an LED attached to it. The implementation may update the indicator asynchronously or synchronously, but in either case it must return quickly. It is initially called with the argument %ETHTOOL_ID_ACTIVE, and must either activate asynchronous updates and return zero, return a negative error or return a positive frequency for synchronous indication (e.g. 1 for one on/off cycle per second). If it returns a frequency then it will be called again at intervals with the argument %ETHTOOL_ID_ON or %ETHTOOL_ID_OFF and should set the state of the indicator accordingly. Finally, it is called with the argument %ETHTOOL_ID_INACTIVE and must deactivate the indicator. Returns a negative error code or zero. @get_ethtool_stats: Return extended statistics about the device. This is only useful if the device maintains statistics not included in &struct rtnl_link_stats64. @begin: Function to be called before any other operation. Returns a negative error code or zero. @complete: Function to be called after any other operation except @begin. Will be called even if the other operation failed. @get_priv_flags: Report driver-specific feature flags. @set_priv_flags: Set driver-specific feature flags. Returns a negative error code or zero. @get_sset_count: Get number of strings that @get_strings will write. @get_rxnfc: Get RX flow classification rules. Returns a negative error code or zero. @set_rxnfc: Set RX flow classification rules. Returns a negative error code or zero. @flash_device: Write a firmware image to device's flash memory. Returns a negative error code or zero. @reset: Reset (part of) the device, as specified by a bitmask of flags from &enum ethtool_reset_flags. Returns a negative error code or zero. @get_rxfh_key_size: Get the size of the RX flow hash key. Returns zero if not supported for this specific device. @get_rxfh_indir_size: Get the size of the RX flow hash indirection table. Returns zero if not supported for this specific device. @get_rxfh: Get the contents of the RX flow hash indirection table, hash key and/or hash function. Returns a negative error code or zero. @set_rxfh: Set the contents of the RX flow hash indirection table, hash key, and/or hash function. Arguments which are set to %NULL or zero will remain unchanged. Returns a negative error code or zero. An error code must be returned if at least one unsupported change was requested. @get_rxfh_context: Get the contents of the RX flow hash indirection table, hash key, and/or hash function assiciated to the given rss context. Returns a negative error code or zero. @set_rxfh_context: Create, remove and configure RSS contexts. Allows setting the contents of the RX flow hash indirection table, hash key, and/or hash function associated to the given context. Arguments which are set to %NULL or zero will remain unchanged. Returns a negative error code or zero. An error code must be returned if at least one unsupported change was requested. @get_channels: Get number of channels. @set_channels: Set number of channels. Returns a negative error code or zero. @get_dump_flag: Get dump flag indicating current dump length, version, and flag of the device. @get_dump_data: Get dump data. @set_dump: Set dump specific flags to the device. @get_ts_info: Get the time stamping and PTP hardware clock capabilities. Drivers supporting transmit time stamps in software should set this to ethtool_op_get_ts_info(). @get_module_info: Get the size and type of the eeprom contained within a plug-in module. @get_module_eeprom: Get the eeprom information from the plug-in module @get_eee: Get Energy-Efficient (EEE) supported and status. @set_eee: Set EEE status (enable/disable) as well as LPI timers. @get_per_queue_coalesce: Get interrupt coalescing parameters per queue. It must check that the given queue number is valid. If neither a RX nor a TX queue has this number, return -EINVAL. If only a RX queue or a TX queue has this number, set the inapplicable fields to ~0 and return 0. Returns a negative error code or zero. @set_per_queue_coalesce: Set interrupt coalescing parameters per queue. It must check that the given queue number is valid. If neither a RX nor a TX queue has this number, return -EINVAL. If only a RX queue or a TX queue has this number, ignore the inapplicable fields. Returns a negative error code or zero. @get_link_ksettings: Get various device settings including Ethernet link settings. The %cmd and %link_mode_masks_nwords fields should be ignored (use %__ETHTOOL_LINK_MODE_MASK_NBITS instead of the latter), any change to them will be overwritten by kernel. Returns a negative error code or zero. @set_link_ksettings: Set various device settings including Ethernet link settings. The %cmd and %link_mode_masks_nwords fields should be ignored (use %__ETHTOOL_LINK_MODE_MASK_NBITS instead of the latter), any change to them will be overwritten by kernel. Returns a negative error code or zero. @get_fecparam: Get the network device Forward Error Correction parameters. @set_fecparam: Set the network device Forward Error Correction parameters. @get_ethtool_phy_stats: Return extended statistics about the PHY device. This is only useful if the device maintains PHY statistics and cannot use the standard PHY library helpers.

All operations are optional (i.e. the function pointer may be set to %NULL) and callers must take this into account. Callers must hold the RTNL lock.

See the structures used by these operations for further documentation. Note that for all operations using a structure ending with a zero- length array, the array is allocated separately in the kernel and is passed to the driver as an additional parameter.

See &struct net_device and &struct net_device_ops for documentation of the generic netdev features interface.

Fields
get_drvinfo: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_drvinfo)>get_regs_len: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>get_regs: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_regs, arg3: *mut c_void)>get_wol: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_wolinfo)>set_wol: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_wolinfo) -> c_int>get_msglevel: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>set_msglevel: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32)>nway_reset: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>get_link: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>get_eeprom_len: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>get_eeprom: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eeprom, arg3: *mut _cargo_bpf_u8) -> c_int>set_eeprom: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eeprom, arg3: *mut _cargo_bpf_u8) -> c_int>get_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_coalesce) -> c_int>set_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_coalesce) -> c_int>get_ringparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_ringparam)>set_ringparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_ringparam) -> c_int>get_pauseparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_pauseparam)>set_pauseparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_pauseparam) -> c_int>self_test: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_test, arg3: *mut _cargo_bpf_u64)>get_strings: Option<unsafe extern "C" fn(arg1: *mut net_device, stringset: _cargo_bpf_u32, arg2: *mut _cargo_bpf_u8)>set_phys_id: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: ethtool_phys_id_state) -> c_int>get_ethtool_stats: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_stats, arg3: *mut _cargo_bpf_u64)>begin: Option<unsafe extern "C" fn(arg1: *mut net_device) -> c_int>complete: Option<unsafe extern "C" fn(arg1: *mut net_device)>get_priv_flags: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>set_priv_flags: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32) -> c_int>get_sset_count: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: c_int) -> c_int>get_rxnfc: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_rxnfc, rule_locs: *mut _cargo_bpf_u32) -> c_int>set_rxnfc: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_rxnfc) -> c_int>flash_device: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_flash) -> c_int>reset: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut _cargo_bpf_u32) -> c_int>get_rxfh_key_size: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>get_rxfh_indir_size: Option<unsafe extern "C" fn(arg1: *mut net_device) -> _cargo_bpf_u32>get_rxfh: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *mut _cargo_bpf_u32, key: *mut _cargo_bpf_u8, hfunc: *mut _cargo_bpf_u8) -> c_int>set_rxfh: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *const _cargo_bpf_u32, key: *const _cargo_bpf_u8, hfunc: _cargo_bpf_u8) -> c_int>get_rxfh_context: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *mut _cargo_bpf_u32, key: *mut _cargo_bpf_u8, hfunc: *mut _cargo_bpf_u8, rss_context: _cargo_bpf_u32) -> c_int>set_rxfh_context: Option<unsafe extern "C" fn(arg1: *mut net_device, indir: *const _cargo_bpf_u32, key: *const _cargo_bpf_u8, hfunc: _cargo_bpf_u8, rss_context: *mut _cargo_bpf_u32, delete: bool_) -> c_int>get_channels: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_channels)>set_channels: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_channels) -> c_int>get_dump_flag: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_dump) -> c_int>get_dump_data: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_dump, arg3: *mut c_void) -> c_int>set_dump: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_dump) -> c_int>get_ts_info: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_ts_info) -> c_int>get_module_info: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_modinfo) -> c_int>get_module_eeprom: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eeprom, arg3: *mut _cargo_bpf_u8) -> c_int>get_eee: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eee) -> c_int>set_eee: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_eee) -> c_int>get_tunable: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *const ethtool_tunable, arg3: *mut c_void) -> c_int>set_tunable: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *const ethtool_tunable, arg3: *const c_void) -> c_int>get_per_queue_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32, arg3: *mut ethtool_coalesce) -> c_int>set_per_queue_coalesce: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: _cargo_bpf_u32, arg3: *mut ethtool_coalesce) -> c_int>get_link_ksettings: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_link_ksettings) -> c_int>set_link_ksettings: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *const ethtool_link_ksettings) -> c_int>get_fecparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_fecparam) -> c_int>set_fecparam: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_fecparam) -> c_int>get_ethtool_phy_stats: Option<unsafe extern "C" fn(arg1: *mut net_device, arg2: *mut ethtool_stats, arg3: *mut _cargo_bpf_u64)>