Struct redbpf_probes::bindings::fwnode_operations

#[repr(C)]pub struct fwnode_operations {
    pub get: Option<unsafe extern "C" fn(fwnode: *mut fwnode_handle) -> *mut fwnode_handle>,
    pub put: Option<unsafe extern "C" fn(fwnode: *mut fwnode_handle)>,
    pub device_is_available: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle) -> bool_>,
    pub device_get_match_data: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, dev: *const device) -> *const c_void>,
    pub property_present: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, propname: *const c_char) -> bool_>,
    pub property_read_int_array: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, propname: *const c_char, elem_size: c_uint, val: *mut c_void, nval: size_t) -> c_int>,
    pub property_read_string_array: Option<unsafe extern "C" fn(fwnode_handle: *const fwnode_handle, propname: *const c_char, val: *mut *const c_char, nval: size_t) -> c_int>,
    pub get_parent: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle) -> *mut fwnode_handle>,
    pub get_next_child_node: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, child: *mut fwnode_handle) -> *mut fwnode_handle>,
    pub get_named_child_node: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, name: *const c_char) -> *mut fwnode_handle>,
    pub get_reference_args: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, prop: *const c_char, nargs_prop: *const c_char, nargs: c_uint, index: c_uint, args: *mut fwnode_reference_args) -> c_int>,
    pub graph_get_next_endpoint: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, prev: *mut fwnode_handle) -> *mut fwnode_handle>,
    pub graph_get_remote_endpoint: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle) -> *mut fwnode_handle>,
    pub graph_get_port_parent: Option<unsafe extern "C" fn(fwnode: *mut fwnode_handle) -> *mut fwnode_handle>,
    pub graph_parse_endpoint: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, endpoint: *mut fwnode_endpoint) -> c_int>,
}

struct fwnode_operations - Operations for fwnode interface @get: Get a reference to an fwnode. @put: Put a reference to an fwnode. @device_get_match_data: Return the device driver match data. @property_present: Return true if a property is present. @property_read_integer_array: Read an array of integer properties. Return zero on success, a negative error code otherwise. @property_read_string_array: Read an array of string properties. Return zero on success, a negative error code otherwise. @get_parent: Return the parent of an fwnode. @get_next_child_node: Return the next child node in an iteration. @get_named_child_node: Return a child node with a given name. @get_reference_args: Return a reference pointed to by a property, with args @graph_get_next_endpoint: Return an endpoint node in an iteration. @graph_get_remote_endpoint: Return the remote endpoint node of a local endpoint node. @graph_get_port_parent: Return the parent node of a port node. @graph_parse_endpoint: Parse endpoint for port and endpoint id.

Fields
get: Option<unsafe extern "C" fn(fwnode: *mut fwnode_handle) -> *mut fwnode_handle>put: Option<unsafe extern "C" fn(fwnode: *mut fwnode_handle)>device_is_available: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle) -> bool_>device_get_match_data: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, dev: *const device) -> *const c_void>property_present: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, propname: *const c_char) -> bool_>property_read_int_array: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, propname: *const c_char, elem_size: c_uint, val: *mut c_void, nval: size_t) -> c_int>property_read_string_array: Option<unsafe extern "C" fn(fwnode_handle: *const fwnode_handle, propname: *const c_char, val: *mut *const c_char, nval: size_t) -> c_int>get_parent: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle) -> *mut fwnode_handle>get_next_child_node: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, child: *mut fwnode_handle) -> *mut fwnode_handle>get_named_child_node: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, name: *const c_char) -> *mut fwnode_handle>get_reference_args: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, prop: *const c_char, nargs_prop: *const c_char, nargs: c_uint, index: c_uint, args: *mut fwnode_reference_args) -> c_int>graph_get_next_endpoint: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, prev: *mut fwnode_handle) -> *mut fwnode_handle>graph_get_remote_endpoint: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle) -> *mut fwnode_handle>graph_get_port_parent: Option<unsafe extern "C" fn(fwnode: *mut fwnode_handle) -> *mut fwnode_handle>graph_parse_endpoint: Option<unsafe extern "C" fn(fwnode: *const fwnode_handle, endpoint: *mut fwnode_endpoint) -> c_int>