Struct redbpf_probes::bindings::hrtimer_cpu_base

#[repr(C)]
pub struct hrtimer_cpu_base {
    pub lock: raw_spinlock_t,
    pub cpu: c_uint,
    pub active_bases: c_uint,
    pub clock_was_set_seq: c_uint,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>,
    pub nr_events: c_uint,
    pub nr_retries: c_ushort,
    pub nr_hangs: c_ushort,
    pub max_hang_time: c_uint,
    pub expires_next: ktime_t,
    pub next_timer: *mut hrtimer,
    pub softirq_expires_next: ktime_t,
    pub softirq_next_timer: *mut hrtimer,
    pub clock_base: [hrtimer_clock_base; 8],
}

struct hrtimer_cpu_base - the per cpu clock bases @lock: lock protecting the base and associated clock bases and timers @cpu: cpu number @active_bases: Bitfield to mark bases with active timers @clock_was_set_seq: Sequence counter of clock was set events @hres_active: State of high resolution mode @in_hrtirq: hrtimer_interrupt() is currently executing @hang_detected: The last hrtimer interrupt detected a hang @softirq_activated: displays, if the softirq is raised - update of softirq related settings is not required then. @nr_events: Total number of hrtimer interrupt events @nr_retries: Total number of hrtimer interrupt retries @nr_hangs: Total number of hrtimer interrupt hangs @max_hang_time: Maximum time spent in hrtimer_interrupt @expires_next: absolute time of the next event, is required for remote hrtimer enqueue; it is the total first expiry time (hard and soft hrtimer are taken into account) @next_timer: Pointer to the first expiring timer @softirq_expires_next: Time to check, if soft queues needs also to be expired @softirq_next_timer: Pointer to the first expiring softirq based timer @clock_base: array of clock bases for this cpu

Note: next_timer is just an optimization for __remove_hrtimer(). Do not dereference the pointer because it is not reliable on cross cpu removals.

Fields
lock: raw_spinlock_tcpu: c_uintactive_bases: c_uintclock_was_set_seq: c_uint_bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>nr_events: c_uintnr_retries: c_ushortnr_hangs: c_ushortmax_hang_time: c_uintexpires_next: ktime_tnext_timer: *mut hrtimersoftirq_expires_next: ktime_tsoftirq_next_timer: *mut hrtimerclock_base: [hrtimer_clock_base; 8]
Methods

impl hrtimer_cpu_base

pub fn hres_active(&self) -> c_uint

pub fn set_hres_active(&mut self, val: c_uint)

pub fn in_hrtirq(&self) -> c_uint

pub fn set_in_hrtirq(&mut self, val: c_uint)

pub fn hang_detected(&self) -> c_uint

pub fn set_hang_detected(&mut self, val: c_uint)

pub fn softirq_activated(&self) -> c_uint

pub fn set_softirq_activated(&mut self, val: c_uint)

pub fn new_bitfield_1(
hres_active: c_uint,
in_hrtirq: c_uint,
hang_detected: c_uint,
softirq_activated: c_uint
) -> __BindgenBitfieldUnit<[u8; 1], u8>