Struct redbpf_probes::bindings::rtnl_link_ops

#[repr(C)]
pub struct rtnl_link_ops {
    pub list: list_head,
    pub kind: *const c_char,
    pub priv_size: usize,
    pub setup: Option<unsafe extern "C" fn(dev: *mut net_device)>,
    pub maxtype: c_int,
    pub policy: *const nla_policy,
    pub validate: Option<unsafe extern "C" fn(tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>,
    pub newlink: Option<unsafe extern "C" fn(src_net: *mut net, dev: *mut net_device, tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>,
    pub changelink: Option<unsafe extern "C" fn(dev: *mut net_device, tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>,
    pub dellink: Option<unsafe extern "C" fn(dev: *mut net_device, head: *mut list_head)>,
    pub get_size: Option<unsafe extern "C" fn(dev: *const net_device) -> usize>,
    pub fill_info: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device) -> c_int>,
    pub get_xstats_size: Option<unsafe extern "C" fn(dev: *const net_device) -> usize>,
    pub fill_xstats: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device) -> c_int>,
    pub get_num_tx_queues: Option<unsafe extern "C" fn() -> c_uint>,
    pub get_num_rx_queues: Option<unsafe extern "C" fn() -> c_uint>,
    pub slave_maxtype: c_int,
    pub slave_policy: *const nla_policy,
    pub slave_changelink: Option<unsafe extern "C" fn(dev: *mut net_device, slave_dev: *mut net_device, tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>,
    pub get_slave_size: Option<unsafe extern "C" fn(dev: *const net_device, slave_dev: *const net_device) -> usize>,
    pub fill_slave_info: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device, slave_dev: *const net_device) -> c_int>,
    pub get_link_net: Option<unsafe extern "C" fn(dev: *const net_device) -> *mut net>,
    pub get_linkxstats_size: Option<unsafe extern "C" fn(dev: *const net_device, attr: c_int) -> usize>,
    pub fill_linkxstats: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device, prividx: *mut c_int, attr: c_int) -> c_int>,
}

struct rtnl_link_ops - rtnetlink link operations

@list: Used internally @kind: Identifier @maxtype: Highest device specific netlink attribute number @policy: Netlink policy for device specific attribute validation @validate: Optional validation function for netlink/changelink parameters @priv_size: sizeof net_device private space @setup: net_device setup function @newlink: Function for configuring and registering a new device @changelink: Function for changing parameters of an existing device @dellink: Function to remove a device @get_size: Function to calculate required room for dumping device specific netlink attributes @fill_info: Function to dump device specific netlink attributes @get_xstats_size: Function to calculate required room for dumping device specific statistics @fill_xstats: Function to dump device specific statistics @get_num_tx_queues: Function to determine number of transmit queues to create when creating a new device. @get_num_rx_queues: Function to determine number of receive queues to create when creating a new device. @get_link_net: Function to get the i/o netns of the device @get_linkxstats_size: Function to calculate the required room for dumping device-specific extended link stats @fill_linkxstats: Function to dump device-specific extended link stats

Fields
list: list_headkind: *const c_charpriv_size: usizesetup: Option<unsafe extern "C" fn(dev: *mut net_device)>maxtype: c_intpolicy: *const nla_policyvalidate: Option<unsafe extern "C" fn(tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>newlink: Option<unsafe extern "C" fn(src_net: *mut net, dev: *mut net_device, tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>changelink: Option<unsafe extern "C" fn(dev: *mut net_device, tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>dellink: Option<unsafe extern "C" fn(dev: *mut net_device, head: *mut list_head)>get_size: Option<unsafe extern "C" fn(dev: *const net_device) -> usize>fill_info: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device) -> c_int>get_xstats_size: Option<unsafe extern "C" fn(dev: *const net_device) -> usize>fill_xstats: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device) -> c_int>get_num_tx_queues: Option<unsafe extern "C" fn() -> c_uint>get_num_rx_queues: Option<unsafe extern "C" fn() -> c_uint>slave_maxtype: c_intslave_policy: *const nla_policyslave_changelink: Option<unsafe extern "C" fn(dev: *mut net_device, slave_dev: *mut net_device, tb: *mut *mut nlattr, data: *mut *mut nlattr, extack: *mut netlink_ext_ack) -> c_int>get_slave_size: Option<unsafe extern "C" fn(dev: *const net_device, slave_dev: *const net_device) -> usize>fill_slave_info: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device, slave_dev: *const net_device) -> c_int>get_link_net: Option<unsafe extern "C" fn(dev: *const net_device) -> *mut net>get_linkxstats_size: Option<unsafe extern "C" fn(dev: *const net_device, attr: c_int) -> usize>fill_linkxstats: Option<unsafe extern "C" fn(skb: *mut sk_buff, dev: *const net_device, prividx: *mut c_int, attr: c_int) -> c_int>