Struct redbpf_probes::bindings::sk_buff

#[repr(C)]pub struct sk_buff {
    pub __bindgen_anon_1: sk_buff__bindgen_ty_1,
    pub __bindgen_anon_2: sk_buff__bindgen_ty_2,
    pub __bindgen_anon_3: sk_buff__bindgen_ty_3,
    pub cb: [c_char; 48],
    pub __bindgen_anon_4: sk_buff__bindgen_ty_4,
    pub _nfct: c_ulong,
    pub len: c_uint,
    pub data_len: c_uint,
    pub mac_len: __u16,
    pub hdr_len: __u16,
    pub queue_mapping: __u16,
    pub __cloned_offset: __IncompleteArrayField<__u8>,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>,
    pub active_extensions: __u8,
    pub headers_start: __IncompleteArrayField<__u32>,
    pub __pkt_type_offset: __IncompleteArrayField<__u8>,
    pub _bitfield_2: __BindgenBitfieldUnit<[u8; 2], u8>,
    pub __pkt_vlan_present_offset: __IncompleteArrayField<__u8>,
    pub _bitfield_3: __BindgenBitfieldUnit<[u8; 3], u8>,
    pub tc_index: __u16,
    pub __bindgen_anon_5: sk_buff__bindgen_ty_5,
    pub priority: __u32,
    pub skb_iif: c_int,
    pub hash: __u32,
    pub vlan_proto: __be16,
    pub vlan_tci: __u16,
    pub __bindgen_anon_6: sk_buff__bindgen_ty_6,
    pub secmark: __u32,
    pub __bindgen_anon_7: sk_buff__bindgen_ty_7,
    pub __bindgen_anon_8: sk_buff__bindgen_ty_8,
    pub inner_transport_header: __u16,
    pub inner_network_header: __u16,
    pub inner_mac_header: __u16,
    pub protocol: __be16,
    pub transport_header: __u16,
    pub network_header: __u16,
    pub mac_header: __u16,
    pub headers_end: __IncompleteArrayField<__u32>,
    pub tail: sk_buff_data_t,
    pub end: sk_buff_data_t,
    pub head: *mut c_uchar,
    pub data: *mut c_uchar,
    pub truesize: c_uint,
    pub users: refcount_t,
    pub extensions: *mut skb_ext,
}

struct sk_buff - socket buffer @next: Next buffer in list @prev: Previous buffer in list @tstamp: Time we arrived/left @rbnode: RB tree node, alternative to next/prev for netem/tcp @sk: Socket we are owned by @dev: Device we arrived on/are leaving by @cb: Control buffer. Free for use by every layer. Put private vars here @_skb_refdst: destination entry (with norefcount bit) @sp: the security path, used for xfrm @len: Length of actual data @data_len: Data length @mac_len: Length of link layer header @hdr_len: writable header length of cloned skb @csum: Checksum (must include start/offset pair) @csum_start: Offset from skb->head where checksumming should start @csum_offset: Offset from csum_start where checksum should be stored @priority: Packet queueing priority @ignore_df: allow local fragmentation @cloned: Head may be cloned (check refcnt to be sure) @ip_summed: Driver fed us an IP checksum @nohdr: Payload reference only, must not modify header @pkt_type: Packet class @fclone: skbuff clone status @ipvs_property: skbuff is owned by ipvs @offload_fwd_mark: Packet was L2-forwarded in hardware @offload_l3_fwd_mark: Packet was L3-forwarded in hardware @tc_skip_classify: do not classify packet. set by IFB device @tc_at_ingress: used within tc_classify to distinguish in/egress @redirected: packet was redirected by packet classifier @from_ingress: packet was redirected from the ingress path @peeked: this packet has been seen already, so stats have been done for it, don't do them again @nf_trace: netfilter packet trace flag @protocol: Packet protocol from driver @destructor: Destruct function @tcp_tsorted_anchor: list structure for TCP (tp->tsorted_sent_queue) @_nfct: Associated connection, if any (with nfctinfo bits) @nf_bridge: Saved data about a bridged frame - see br_netfilter.c @skb_iif: ifindex of device we arrived on @tc_index: Traffic control index @hash: the packet hash @queue_mapping: Queue mapping for multiqueue devices @pfmemalloc: skbuff was allocated from PFMEMALLOC reserves @active_extensions: active extensions (skb_ext_id types) @ndisc_nodetype: router type (from link layer) @ooo_okay: allow the mapping of a socket to a queue to be changed @l4_hash: indicate hash is a canonical 4-tuple hash over transport ports. @sw_hash: indicates hash was computed in software stack @wifi_acked_valid: wifi_acked was set @wifi_acked: whether frame was acked on wifi or not @no_fcs: Request NIC to treat last 4 bytes as Ethernet FCS @csum_not_inet: use CRC32c to resolve CHECKSUM_PARTIAL @dst_pending_confirm: need to confirm neighbour @decrypted: Decrypted SKB @napi_id: id of the NAPI struct this skb came from @secmark: security marking @mark: Generic packet mark @vlan_proto: vlan encapsulation protocol @vlan_tci: vlan tag control information @inner_protocol: Protocol (encapsulation) @inner_transport_header: Inner transport layer header (encapsulation) @inner_network_header: Network layer header (encapsulation) @inner_mac_header: Link layer header (encapsulation) @transport_header: Transport layer header @network_header: Network layer header @mac_header: Link layer header @tail: Tail pointer @end: End pointer @head: Head of buffer @data: Data head pointer @truesize: Buffer size @users: User count - see {datagram,tcp}.c @extensions: allocated extensions, valid if active_extensions is nonzero

Fields
__bindgen_anon_1: sk_buff__bindgen_ty_1__bindgen_anon_2: sk_buff__bindgen_ty_2__bindgen_anon_3: sk_buff__bindgen_ty_3cb: [c_char; 48]__bindgen_anon_4: sk_buff__bindgen_ty_4_nfct: c_ulonglen: c_uintdata_len: c_uintmac_len: __u16hdr_len: __u16queue_mapping: __u16__cloned_offset: __IncompleteArrayField<__u8>_bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>active_extensions: __u8headers_start: __IncompleteArrayField<__u32>__pkt_type_offset: __IncompleteArrayField<__u8>_bitfield_2: __BindgenBitfieldUnit<[u8; 2], u8>__pkt_vlan_present_offset: __IncompleteArrayField<__u8>_bitfield_3: __BindgenBitfieldUnit<[u8; 3], u8>tc_index: __u16__bindgen_anon_5: sk_buff__bindgen_ty_5priority: __u32skb_iif: c_inthash: __u32vlan_proto: __be16vlan_tci: __u16__bindgen_anon_6: sk_buff__bindgen_ty_6secmark: __u32__bindgen_anon_7: sk_buff__bindgen_ty_7__bindgen_anon_8: sk_buff__bindgen_ty_8inner_transport_header: __u16inner_network_header: __u16inner_mac_header: __u16protocol: __be16transport_header: __u16network_header: __u16mac_header: __u16headers_end: __IncompleteArrayField<__u32>tail: sk_buff_data_tend: sk_buff_data_thead: *mut c_uchardata: *mut c_uchartruesize: c_uintusers: refcount_textensions: *mut skb_ext
Implementations

impl sk_buff

pub fn cloned(&self) -> __u8

pub fn set_cloned(&mut self, val: __u8)

pub fn nohdr(&self) -> __u8

pub fn set_nohdr(&mut self, val: __u8)

pub fn fclone(&self) -> __u8

pub fn set_fclone(&mut self, val: __u8)

pub fn peeked(&self) -> __u8

pub fn set_peeked(&mut self, val: __u8)

pub fn head_frag(&self) -> __u8

pub fn set_head_frag(&mut self, val: __u8)

pub fn pfmemalloc(&self) -> __u8

pub fn set_pfmemalloc(&mut self, val: __u8)

pub fn new_bitfield_1(
cloned: __u8,
nohdr: __u8,
fclone: __u8,
peeked: __u8,
head_frag: __u8,
pfmemalloc: __u8
) -> __BindgenBitfieldUnit<[u8; 1], u8>

pub fn pkt_type(&self) -> __u8

pub fn set_pkt_type(&mut self, val: __u8)

pub fn ignore_df(&self) -> __u8

pub fn set_ignore_df(&mut self, val: __u8)

pub fn nf_trace(&self) -> __u8

pub fn set_nf_trace(&mut self, val: __u8)

pub fn ip_summed(&self) -> __u8

pub fn set_ip_summed(&mut self, val: __u8)

pub fn ooo_okay(&self) -> __u8

pub fn set_ooo_okay(&mut self, val: __u8)

pub fn l4_hash(&self) -> __u8

pub fn set_l4_hash(&mut self, val: __u8)

pub fn sw_hash(&self) -> __u8

pub fn set_sw_hash(&mut self, val: __u8)

pub fn wifi_acked_valid(&self) -> __u8

pub fn set_wifi_acked_valid(&mut self, val: __u8)

pub fn wifi_acked(&self) -> __u8

pub fn set_wifi_acked(&mut self, val: __u8)

pub fn no_fcs(&self) -> __u8

pub fn set_no_fcs(&mut self, val: __u8)

pub fn encapsulation(&self) -> __u8

pub fn set_encapsulation(&mut self, val: __u8)

pub fn encap_hdr_csum(&self) -> __u8

pub fn set_encap_hdr_csum(&mut self, val: __u8)

pub fn csum_valid(&self) -> __u8

pub fn set_csum_valid(&mut self, val: __u8)

pub fn new_bitfield_2(
pkt_type: __u8,
ignore_df: __u8,
nf_trace: __u8,
ip_summed: __u8,
ooo_okay: __u8,
l4_hash: __u8,
sw_hash: __u8,
wifi_acked_valid: __u8,
wifi_acked: __u8,
no_fcs: __u8,
encapsulation: __u8,
encap_hdr_csum: __u8,
csum_valid: __u8
) -> __BindgenBitfieldUnit<[u8; 2], u8>

pub fn vlan_present(&self) -> __u8

pub fn set_vlan_present(&mut self, val: __u8)

pub fn csum_complete_sw(&self) -> __u8

pub fn set_csum_complete_sw(&mut self, val: __u8)

pub fn csum_level(&self) -> __u8

pub fn set_csum_level(&mut self, val: __u8)

pub fn csum_not_inet(&self) -> __u8

pub fn set_csum_not_inet(&mut self, val: __u8)

pub fn dst_pending_confirm(&self) -> __u8

pub fn set_dst_pending_confirm(&mut self, val: __u8)

pub fn ndisc_nodetype(&self) -> __u8

pub fn set_ndisc_nodetype(&mut self, val: __u8)

pub fn ipvs_property(&self) -> __u8

pub fn set_ipvs_property(&mut self, val: __u8)

pub fn inner_protocol_type(&self) -> __u8

pub fn set_inner_protocol_type(&mut self, val: __u8)

pub fn remcsum_offload(&self) -> __u8

pub fn set_remcsum_offload(&mut self, val: __u8)

pub fn offload_fwd_mark(&self) -> __u8

pub fn set_offload_fwd_mark(&mut self, val: __u8)

pub fn offload_l3_fwd_mark(&self) -> __u8

pub fn set_offload_l3_fwd_mark(&mut self, val: __u8)

pub fn tc_skip_classify(&self) -> __u8

pub fn set_tc_skip_classify(&mut self, val: __u8)

pub fn tc_at_ingress(&self) -> __u8

pub fn set_tc_at_ingress(&mut self, val: __u8)

pub fn redirected(&self) -> __u8

pub fn set_redirected(&mut self, val: __u8)

pub fn from_ingress(&self) -> __u8

pub fn set_from_ingress(&mut self, val: __u8)

pub fn decrypted(&self) -> __u8

pub fn set_decrypted(&mut self, val: __u8)

pub fn new_bitfield_3(
vlan_present: __u8,
csum_complete_sw: __u8,
csum_level: __u8,
csum_not_inet: __u8,
dst_pending_confirm: __u8,
ndisc_nodetype: __u8,
ipvs_property: __u8,
inner_protocol_type: __u8,
remcsum_offload: __u8,
offload_fwd_mark: __u8,
offload_l3_fwd_mark: __u8,
tc_skip_classify: __u8,
tc_at_ingress: __u8,
redirected: __u8,
from_ingress: __u8,
decrypted: __u8
) -> __BindgenBitfieldUnit<[u8; 3], u8>