Struct redbpf_probes::bindings::sock_common

#[repr(C)]
pub struct sock_common {
    pub __bindgen_anon_1: sock_common__bindgen_ty_1,
    pub __bindgen_anon_2: sock_common__bindgen_ty_2,
    pub __bindgen_anon_3: sock_common__bindgen_ty_3,
    pub skc_family: c_ushort,
    pub skc_state: c_uchar,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>,
    pub skc_bound_dev_if: c_int,
    pub __bindgen_anon_4: sock_common__bindgen_ty_4,
    pub skc_prot: *mut proto,
    pub skc_net: possible_net_t,
    pub skc_v6_daddr: in6_addr,
    pub skc_v6_rcv_saddr: in6_addr,
    pub skc_cookie: atomic64_t,
    pub __bindgen_anon_5: sock_common__bindgen_ty_5,
    pub skc_dontcopy_begin: __IncompleteArrayField<c_int>,
    pub __bindgen_anon_6: sock_common__bindgen_ty_6,
    pub skc_tx_queue_mapping: c_int,
    pub __bindgen_anon_7: sock_common__bindgen_ty_7,
    pub skc_refcnt: refcount_t,
    pub skc_dontcopy_end: __IncompleteArrayField<c_int>,
    pub __bindgen_anon_8: sock_common__bindgen_ty_8,
}

struct sock_common - minimal network layer representation of sockets @skc_daddr: Foreign IPv4 addr @skc_rcv_saddr: Bound local IPv4 addr @skc_hash: hash value used with various protocol lookup tables @skc_u16hashes: two u16 hash values used by UDP lookup tables @skc_dport: placeholder for inet_dport/tw_dport @skc_num: placeholder for inet_num/tw_num @skc_family: network address family @skc_state: Connection state @skc_reuse: %SO_REUSEADDR setting @skc_reuseport: %SO_REUSEPORT setting @skc_bound_dev_if: bound device index if != 0 @skc_bind_node: bind hash linkage for various protocol lookup tables @skc_portaddr_node: second hash linkage for UDP/UDP-Lite protocol @skc_prot: protocol handlers inside a network family @skc_net: reference to the network namespace of this socket @skc_node: main hash linkage for various protocol lookup tables @skc_nulls_node: main hash linkage for TCP/UDP/UDP-Lite protocol @skc_tx_queue_mapping: tx queue number for this connection @skc_flags: place holder for sk_flags %SO_LINGER (l_onoff), %SO_BROADCAST, %SO_KEEPALIVE, %SO_OOBINLINE settings, %SO_TIMESTAMPING settings @skc_incoming_cpu: record/match cpu processing incoming packets @skc_refcnt: reference count

This is the minimal network layer representation of sockets, the header for struct sock and struct inet_timewait_sock.

Fields
__bindgen_anon_1: sock_common__bindgen_ty_1__bindgen_anon_2: sock_common__bindgen_ty_2__bindgen_anon_3: sock_common__bindgen_ty_3skc_family: c_ushortskc_state: c_uchar_bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>skc_bound_dev_if: c_int__bindgen_anon_4: sock_common__bindgen_ty_4skc_prot: *mut protoskc_net: possible_net_tskc_v6_daddr: in6_addrskc_v6_rcv_saddr: in6_addrskc_cookie: atomic64_t__bindgen_anon_5: sock_common__bindgen_ty_5skc_dontcopy_begin: __IncompleteArrayField<c_int>__bindgen_anon_6: sock_common__bindgen_ty_6skc_tx_queue_mapping: c_int__bindgen_anon_7: sock_common__bindgen_ty_7skc_refcnt: refcount_tskc_dontcopy_end: __IncompleteArrayField<c_int>__bindgen_anon_8: sock_common__bindgen_ty_8
Methods

impl sock_common

pub fn skc_reuse(&self) -> c_uchar

pub fn set_skc_reuse(&mut self, val: c_uchar)

pub fn skc_reuseport(&self) -> c_uchar

pub fn set_skc_reuseport(&mut self, val: c_uchar)

pub fn skc_ipv6only(&self) -> c_uchar

pub fn set_skc_ipv6only(&mut self, val: c_uchar)

pub fn skc_net_refcnt(&self) -> c_uchar

pub fn set_skc_net_refcnt(&mut self, val: c_uchar)

pub fn new_bitfield_1(
skc_reuse: c_uchar,
skc_reuseport: c_uchar,
skc_ipv6only: c_uchar,
skc_net_refcnt: c_uchar
) -> __BindgenBitfieldUnit<[u8; 1], u8>