Struct redbpf_probes::bindings::wakeup_source

#[repr(C)]
pub struct wakeup_source {
    pub name: *const c_char,
    pub entry: list_head,
    pub lock: spinlock_t,
    pub wakeirq: *mut wake_irq,
    pub timer: timer_list,
    pub timer_expires: c_ulong,
    pub total_time: ktime_t,
    pub max_time: ktime_t,
    pub last_time: ktime_t,
    pub start_prevent_time: ktime_t,
    pub prevent_sleep_time: ktime_t,
    pub event_count: c_ulong,
    pub active_count: c_ulong,
    pub relax_count: c_ulong,
    pub expire_count: c_ulong,
    pub wakeup_count: c_ulong,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>,
    pub __bindgen_padding_0: [u8; 7],
}

struct wakeup_source - Representation of wakeup sources

@name: Name of the wakeup source @entry: Wakeup source list entry @lock: Wakeup source lock @wakeirq: Optional device specific wakeirq @timer: Wakeup timer list @timer_expires: Wakeup timer expiration @total_time: Total time this wakeup source has been active. @max_time: Maximum time this wakeup source has been continuously active. @last_time: Monotonic clock when the wakeup source's was touched last time. @prevent_sleep_time: Total time this source has been preventing autosleep. @event_count: Number of signaled wakeup events. @active_count: Number of times the wakeup source was activated. @relax_count: Number of times the wakeup source was deactivated. @expire_count: Number of times the wakeup source's timeout has expired. @wakeup_count: Number of times the wakeup source might abort suspend. @active: Status of the wakeup source. @has_timeout: The wakeup source has been activated with a timeout.

Fields
name: *const c_charentry: list_headlock: spinlock_twakeirq: *mut wake_irqtimer: timer_listtimer_expires: c_ulongtotal_time: ktime_tmax_time: ktime_tlast_time: ktime_tstart_prevent_time: ktime_tprevent_sleep_time: ktime_tevent_count: c_ulongactive_count: c_ulongrelax_count: c_ulongexpire_count: c_ulongwakeup_count: c_ulong_bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>__bindgen_padding_0: [u8; 7]
Methods

impl wakeup_source

pub fn active(&self) -> bool_

pub fn set_active(&mut self, val: bool_)

pub fn autosleep_enabled(&self) -> bool_

pub fn set_autosleep_enabled(&mut self, val: bool_)

pub fn new_bitfield_1(
active: bool_,
autosleep_enabled: bool_
) -> __BindgenBitfieldUnit<[u8; 1], u8>