Module redbpf_probes::helpers

Wrappers for the helper functions provided by the BPF subsystem.

See http://man7.org/linux/man-pages/man7/bpf-helpers.7.html.

Re-exports

pub use ufmt;

pub use gen::*;

Modules

gen
Low-level, generated bindings for the helper functions provided by the BPF subsystem.

Structs

TraceMessage

Functions

bpf_get_current_comm
Returns the comm attribute of the current task. The comm attribute contains the name of the executable (excluding the path) for the current task.

bpf_get_current_pid_tgid
Returns the current tgid and pid, encoded as current_task->tgid << 32 | current_task->pid.

bpf_get_current_uid_gid
Returns the current GID and UID, encoded as current_gid << 32 | current_uid.

bpf_ktime_get_ns
Returns the time elapsed since system boot, in nanoseconds.

bpf_perf_event_output

bpf_probe_read⚠

bpf_trace_printk