Module redbpf_probes::uprobe::prelude

The Kprobe Prelude

The purpose of this module is to alleviate imports of the common kprobe types by adding a glob import to the top of kprobe programs:

use redbpf_probes::kprobe::prelude::*;

Re-exports

pub use crate::bindings::*;

pub use crate::helpers::*;

pub use crate::maps::*;

pub use crate::kprobe::*;

Macros

program
Generates program metadata.

Type Definitions

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

c_void

int8_t

int16_t

int32_t

int64_t

intmax_t

intptr_t

ptrdiff_t

size_t

ssize_t

uint8_t

uint16_t

uint32_t

uint64_t

uintmax_t

uintptr_t

Attribute Macros

map
Attribute macro that must be used when creating eBPF maps.

uprobe
Attribute macro that must be used to define uprobes.

uretprobe
Attribute macro that must be used to define uretprobes.