Struct redbpf_probes::bindings::net_device

#[repr(C)]
pub struct net_device {
    pub name: [c_char; 16],
    pub name_hlist: hlist_node,
    pub ifalias: *mut dev_ifalias,
    pub mem_end: c_ulong,
    pub mem_start: c_ulong,
    pub base_addr: c_ulong,
    pub irq: c_int,
    pub state: c_ulong,
    pub dev_list: list_head,
    pub napi_list: list_head,
    pub unreg_list: list_head,
    pub close_list: list_head,
    pub ptype_all: list_head,
    pub ptype_specific: list_head,
    pub adj_list: net_device__bindgen_ty_1,
    pub features: netdev_features_t,
    pub hw_features: netdev_features_t,
    pub wanted_features: netdev_features_t,
    pub vlan_features: netdev_features_t,
    pub hw_enc_features: netdev_features_t,
    pub mpls_features: netdev_features_t,
    pub gso_partial_features: netdev_features_t,
    pub ifindex: c_int,
    pub group: c_int,
    pub stats: net_device_stats,
    pub rx_dropped: atomic_long_t,
    pub tx_dropped: atomic_long_t,
    pub rx_nohandler: atomic_long_t,
    pub carrier_up_count: atomic_t,
    pub carrier_down_count: atomic_t,
    pub wireless_handlers: *mut iw_handler_def,
    pub wireless_data: *mut iw_public_data,
    pub netdev_ops: *const net_device_ops,
    pub ethtool_ops: *const ethtool_ops,
    pub ndisc_ops: *mut ndisc_ops,
    pub xfrmdev_ops: *const xfrmdev_ops,
    pub header_ops: *const header_ops,
    pub flags: c_uint,
    pub priv_flags: c_uint,
    pub gflags: c_ushort,
    pub padded: c_ushort,
    pub operstate: c_uchar,
    pub link_mode: c_uchar,
    pub if_port: c_uchar,
    pub dma: c_uchar,
    pub mtu: c_uint,
    pub min_mtu: c_uint,
    pub max_mtu: c_uint,
    pub type_: c_ushort,
    pub hard_header_len: c_ushort,
    pub min_header_len: c_uchar,
    pub needed_headroom: c_ushort,
    pub needed_tailroom: c_ushort,
    pub perm_addr: [c_uchar; 32],
    pub addr_assign_type: c_uchar,
    pub addr_len: c_uchar,
    pub neigh_priv_len: c_ushort,
    pub dev_id: c_ushort,
    pub dev_port: c_ushort,
    pub addr_list_lock: spinlock_t,
    pub name_assign_type: c_uchar,
    pub uc_promisc: bool_,
    pub uc: netdev_hw_addr_list,
    pub mc: netdev_hw_addr_list,
    pub dev_addrs: netdev_hw_addr_list,
    pub queues_kset: *mut kset,
    pub promiscuity: c_uint,
    pub allmulti: c_uint,
    pub vlan_info: *mut vlan_info,
    pub atalk_ptr: *mut c_void,
    pub ip_ptr: *mut in_device,
    pub dn_ptr: *mut dn_dev,
    pub ip6_ptr: *mut inet6_dev,
    pub ax25_ptr: *mut c_void,
    pub ieee80211_ptr: *mut wireless_dev,
    pub ieee802154_ptr: *mut wpan_dev,
    pub mpls_ptr: *mut mpls_dev,
    pub dev_addr: *mut c_uchar,
    pub _rx: *mut netdev_rx_queue,
    pub num_rx_queues: c_uint,
    pub real_num_rx_queues: c_uint,
    pub xdp_prog: *mut bpf_prog,
    pub gro_flush_timeout: c_ulong,
    pub rx_handler: rx_handler_func_t,
    pub rx_handler_data: *mut c_void,
    pub miniq_ingress: *mut mini_Qdisc,
    pub ingress_queue: *mut netdev_queue,
    pub nf_hooks_ingress: *mut nf_hook_entries,
    pub broadcast: [c_uchar; 32],
    pub rx_cpu_rmap: *mut cpu_rmap,
    pub index_hlist: hlist_node,
    pub __bindgen_padding_0: [u64; 4],
    pub _tx: *mut netdev_queue,
    pub num_tx_queues: c_uint,
    pub real_num_tx_queues: c_uint,
    pub qdisc: *mut Qdisc,
    pub qdisc_hash: [hlist_head; 16],
    pub tx_queue_len: c_uint,
    pub tx_global_lock: spinlock_t,
    pub watchdog_timeo: c_int,
    pub xps_maps: *mut xps_dev_maps,
    pub miniq_egress: *mut mini_Qdisc,
    pub watchdog_timer: timer_list,
    pub pcpu_refcnt: *mut c_int,
    pub todo_list: list_head,
    pub link_watch_list: list_head,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>,
    pub dismantle: bool_,
    pub _bitfield_2: __BindgenBitfieldUnit<[u8; 2], u16>,
    pub needs_free_netdev: bool_,
    pub priv_destructor: Option<unsafe extern "C" fn(dev: *mut net_device)>,
    pub npinfo: *mut netpoll_info,
    pub nd_net: possible_net_t,
    pub __bindgen_anon_1: net_device__bindgen_ty_4,
    pub garp_port: *mut garp_port,
    pub mrp_port: *mut mrp_port,
    pub dev: device,
    pub sysfs_groups: [*const attribute_group; 4],
    pub sysfs_rx_queue_group: *const attribute_group,
    pub rtnl_link_ops: *const rtnl_link_ops,
    pub gso_max_size: c_uint,
    pub gso_max_segs: u16,
    pub dcbnl_ops: *const dcbnl_rtnl_ops,
    pub num_tc: u8,
    pub tc_to_txq: [netdev_tc_txq; 16],
    pub prio_tc_map: [u8; 16],
    pub fcoe_ddp_xid: c_uint,
    pub priomap: *mut netprio_map,
    pub phydev: *mut phy_device,
    pub qdisc_tx_busylock: *mut lock_class_key,
    pub qdisc_running_key: *mut lock_class_key,
    pub proto_down: bool_,
    pub __bindgen_padding_1: [u8; 47],
}

struct net_device - The DEVICE structure.

Actually, this whole structure is a big mistake. It mixes I/O data with strictly "high-level" data, and it has to know about almost every data structure used in the INET module.

@name: This is the first field of the "visible" part of this structure (i.e. as seen by users in the "Space.c" file). It is the name of the interface.

@name_hlist: Device name hash chain, please keep it close to name[] @ifalias: SNMP alias @mem_end: Shared memory end @mem_start: Shared memory start @base_addr: Device I/O address @irq: Device IRQ number

@state: Generic network queuing layer state, see netdev_state_t @dev_list: The global list of network devices @napi_list: List entry used for polling NAPI devices @unreg_list: List entry when we are unregistering the device; see the function unregister_netdev @close_list: List entry used when we are closing the device @ptype_all: Device-specific packet handlers for all protocols @ptype_specific: Device-specific, protocol-specific packet handlers

@adj_list: Directly linked devices, like slaves for bonding @features: Currently active device features @hw_features: User-changeable features

@wanted_features: User-requested features @vlan_features: Mask of features inheritable by VLAN devices

@hw_enc_features: Mask of features inherited by encapsulating devices This field indicates what encapsulation offloads the hardware is capable of doing, and drivers will need to set them appropriately.

@mpls_features: Mask of features inheritable by MPLS

@ifindex: interface index @group: The group the device belongs to

@stats: Statistics struct, which was left as a legacy, use rtnl_link_stats64 instead

@rx_dropped: Dropped packets by core network, do not use this in drivers @tx_dropped: Dropped packets by core network, do not use this in drivers @rx_nohandler: nohandler dropped packets by core network on inactive devices, do not use this in drivers @carrier_up_count: Number of times the carrier has been up @carrier_down_count: Number of times the carrier has been down

@wireless_handlers: List of functions to handle Wireless Extensions, instead of ioctl, see <net/iw_handler.h> for details. @wireless_data: Instance data managed by the core of wireless extensions

@netdev_ops: Includes several pointers to callbacks, if one wants to override the ndo_*() functions @ethtool_ops: Management operations @ndisc_ops: Includes callbacks for different IPv6 neighbour discovery handling. Necessary for e.g. 6LoWPAN. @header_ops: Includes callbacks for creating,parsing,caching,etc of Layer 2 headers.

@flags: Interface flags (a la BSD) @priv_flags: Like 'flags' but invisible to userspace, see if.h for the definitions @gflags: Global flags ( kept as legacy ) @padded: How much padding added by alloc_netdev() @operstate: RFC2863 operstate @link_mode: Mapping policy to operstate @if_port: Selectable AUI, TP, ... @dma: DMA channel @mtu: Interface MTU value @min_mtu: Interface Minimum MTU value @max_mtu: Interface Maximum MTU value @type: Interface hardware type @hard_header_len: Maximum hardware header length. @min_header_len: Minimum hardware header length

@needed_headroom: Extra headroom the hardware may need, but not in all cases can this be guaranteed @needed_tailroom: Extra tailroom the hardware may need, but not in all cases can this be guaranteed. Some cases also use LL_MAX_HEADER instead to allocate the skb

interface address info:

@perm_addr:		Permanent hw address
@addr_assign_type:	Hw address assignment type
@addr_len:		Hardware address length

@neigh_priv_len: Used in neigh_alloc() @dev_id: Used to differentiate devices that share the same link layer address @dev_port: Used to differentiate devices that share the same function @addr_list_lock: XXX: need comments on this one @uc_promisc: Counter that indicates promiscuous mode has been enabled due to the need to listen to additional unicast addresses in a device that does not implement ndo_set_rx_mode() @uc: unicast mac addresses @mc: multicast mac addresses @dev_addrs: list of device hw addresses @queues_kset: Group of all Kobjects in the Tx and RX queues @promiscuity: Number of times the NIC is told to work in promiscuous mode; if it becomes 0 the NIC will exit promiscuous mode @allmulti: Counter, enables or disables allmulticast mode

@vlan_info: VLAN info @dsa_ptr: dsa specific data @tipc_ptr: TIPC specific data @atalk_ptr: AppleTalk link @ip_ptr: IPv4 specific data @dn_ptr: DECnet specific data @ip6_ptr: IPv6 specific data @ax25_ptr: AX.25 specific data @ieee80211_ptr: IEEE 802.11 specific data, assign before registering

@dev_addr: Hw address (before bcast, because most packets are unicast)

@_rx: Array of RX queues @num_rx_queues: Number of RX queues allocated at register_netdev() time @real_num_rx_queues: Number of RX queues currently active in device

@rx_handler: handler for received packets @rx_handler_data: XXX: need comments on this one @miniq_ingress: ingress/clsact qdisc specific data for ingress processing @ingress_queue: XXX: need comments on this one @broadcast: hw bcast address

@rx_cpu_rmap: CPU reverse-mapping for RX completion interrupts, indexed by RX queue number. Assigned by driver. This must only be set if the ndo_rx_flow_steer operation is defined @index_hlist: Device index hash chain

@_tx: Array of TX queues @num_tx_queues: Number of TX queues allocated at alloc_netdev_mq() time @real_num_tx_queues: Number of TX queues currently active in device @qdisc: Root qdisc from userspace point of view @tx_queue_len: Max frames per queue allowed @tx_global_lock: XXX: need comments on this one

@xps_maps: XXX: need comments on this one @miniq_egress: clsact qdisc specific data for egress processing @watchdog_timeo: Represents the timeout that is used by the watchdog (see dev_watchdog()) @watchdog_timer: List of timers

@pcpu_refcnt: Number of references to this device @todo_list: Delayed register/unregister @link_watch_list: XXX: need comments on this one

@reg_state: Register/unregister state machine @dismantle: Device is going to be freed @rtnl_link_state: This enum represents the phases of creating a new link

@needs_free_netdev: Should unregister perform free_netdev? @priv_destructor: Called from unregister @npinfo: XXX: need comments on this one @nd_net: Network namespace this network device is inside

@ml_priv:	Mid-layer private
@lstats:	Loopback statistics
@tstats:	Tunnel statistics
@dstats:	Dummy statistics
@vstats:	Virtual ethernet statistics

@garp_port: GARP @mrp_port: MRP

@dev: Class/net/name entry @sysfs_groups: Space for optional device, statistics and wireless sysfs groups

@sysfs_rx_queue_group: Space for optional per-rx queue attributes @rtnl_link_ops: Rtnl_link_ops

@gso_max_size: Maximum size of generic segmentation offload @gso_max_segs: Maximum number of segments that can be passed to the NIC for GSO

@dcbnl_ops: Data Center Bridging netlink ops @num_tc: Number of traffic classes in the net device @tc_to_txq: XXX: need comments on this one @prio_tc_map: XXX: need comments on this one

@fcoe_ddp_xid: Max exchange id for FCoE LRO by ddp

@priomap: XXX: need comments on this one @phydev: Physical device may attach itself for hardware timestamping

@qdisc_tx_busylock: lockdep class annotating Qdisc->busylock spinlock @qdisc_running_key: lockdep class annotating Qdisc->running seqcount

@proto_down: protocol port state information can be sent to the switch driver and used to set the phys state of the switch port.

FIXME: cleanup struct net_device such that network protocol info moves out.

Fields
name: [c_char; 16]name_hlist: hlist_nodeifalias: *mut dev_ifaliasmem_end: c_ulongmem_start: c_ulongbase_addr: c_ulongirq: c_intstate: c_ulongdev_list: list_headnapi_list: list_headunreg_list: list_headclose_list: list_headptype_all: list_headptype_specific: list_headadj_list: net_device__bindgen_ty_1features: netdev_features_thw_features: netdev_features_twanted_features: netdev_features_tvlan_features: netdev_features_thw_enc_features: netdev_features_tmpls_features: netdev_features_tgso_partial_features: netdev_features_tifindex: c_intgroup: c_intstats: net_device_statsrx_dropped: atomic_long_ttx_dropped: atomic_long_trx_nohandler: atomic_long_tcarrier_up_count: atomic_tcarrier_down_count: atomic_twireless_handlers: *mut iw_handler_defwireless_data: *mut iw_public_datanetdev_ops: *const net_device_opsethtool_ops: *const ethtool_opsndisc_ops: *mut ndisc_opsxfrmdev_ops: *const xfrmdev_opsheader_ops: *const header_opsflags: c_uintpriv_flags: c_uintgflags: c_ushortpadded: c_ushortoperstate: c_ucharlink_mode: c_ucharif_port: c_uchardma: c_ucharmtu: c_uintmin_mtu: c_uintmax_mtu: c_uinttype_: c_ushorthard_header_len: c_ushortmin_header_len: c_ucharneeded_headroom: c_ushortneeded_tailroom: c_ushortperm_addr: [c_uchar; 32]addr_assign_type: c_ucharaddr_len: c_ucharneigh_priv_len: c_ushortdev_id: c_ushortdev_port: c_ushortaddr_list_lock: spinlock_tname_assign_type: c_ucharuc_promisc: bool_uc: netdev_hw_addr_listmc: netdev_hw_addr_listdev_addrs: netdev_hw_addr_listqueues_kset: *mut ksetpromiscuity: c_uintallmulti: c_uintvlan_info: *mut vlan_infoatalk_ptr: *mut c_voidip_ptr: *mut in_devicedn_ptr: *mut dn_devip6_ptr: *mut inet6_devax25_ptr: *mut c_voidieee80211_ptr: *mut wireless_devieee802154_ptr: *mut wpan_devmpls_ptr: *mut mpls_devdev_addr: *mut c_uchar_rx: *mut netdev_rx_queuenum_rx_queues: c_uintreal_num_rx_queues: c_uintxdp_prog: *mut bpf_proggro_flush_timeout: c_ulongrx_handler: rx_handler_func_trx_handler_data: *mut c_voidminiq_ingress: *mut mini_Qdiscingress_queue: *mut netdev_queuenf_hooks_ingress: *mut nf_hook_entriesbroadcast: [c_uchar; 32]rx_cpu_rmap: *mut cpu_rmapindex_hlist: hlist_node__bindgen_padding_0: [u64; 4]_tx: *mut netdev_queuenum_tx_queues: c_uintreal_num_tx_queues: c_uintqdisc: *mut Qdiscqdisc_hash: [hlist_head; 16]tx_queue_len: c_uinttx_global_lock: spinlock_twatchdog_timeo: c_intxps_maps: *mut xps_dev_mapsminiq_egress: *mut mini_Qdiscwatchdog_timer: timer_listpcpu_refcnt: *mut c_inttodo_list: list_headlink_watch_list: list_head_bitfield_1: __BindgenBitfieldUnit<[u8; 1], u8>dismantle: bool__bitfield_2: __BindgenBitfieldUnit<[u8; 2], u16>needs_free_netdev: bool_priv_destructor: Option<unsafe extern "C" fn(dev: *mut net_device)>npinfo: *mut netpoll_infond_net: possible_net_t__bindgen_anon_1: net_device__bindgen_ty_4garp_port: *mut garp_portmrp_port: *mut mrp_portdev: devicesysfs_groups: [*const attribute_group; 4]sysfs_rx_queue_group: *const attribute_grouprtnl_link_ops: *const rtnl_link_opsgso_max_size: c_uintgso_max_segs: u16dcbnl_ops: *const dcbnl_rtnl_opsnum_tc: u8tc_to_txq: [netdev_tc_txq; 16]prio_tc_map: [u8; 16]fcoe_ddp_xid: c_uintpriomap: *mut netprio_mapphydev: *mut phy_deviceqdisc_tx_busylock: *mut lock_class_keyqdisc_running_key: *mut lock_class_keyproto_down: bool___bindgen_padding_1: [u8; 47]
Methods

impl net_device

pub fn reg_state(&self) -> net_device__bindgen_ty_2

pub fn set_reg_state(&mut self, val: net_device__bindgen_ty_2)

pub fn new_bitfield_1(
reg_state: net_device__bindgen_ty_2
) -> __BindgenBitfieldUnit<[u8; 1], u8>

pub fn rtnl_link_state(&self) -> net_device__bindgen_ty_3

pub fn set_rtnl_link_state(&mut self, val: net_device__bindgen_ty_3)

pub fn new_bitfield_2(
rtnl_link_state: net_device__bindgen_ty_3
) -> __BindgenBitfieldUnit<[u8; 2], u16>