Struct redbpf_probes::bindings::sock

#[repr(C)]
pub struct sock {
    pub __sk_common: sock_common,
    pub sk_lock: socket_lock_t,
    pub sk_drops: atomic_t,
    pub sk_rcvlowat: c_int,
    pub sk_error_queue: sk_buff_head,
    pub sk_receive_queue: sk_buff_head,
    pub sk_backlog: sock__bindgen_ty_1,
    pub sk_forward_alloc: c_int,
    pub sk_ll_usec: c_uint,
    pub sk_napi_id: c_uint,
    pub sk_rcvbuf: c_int,
    pub sk_filter: *mut sk_filter,
    pub __bindgen_anon_1: sock__bindgen_ty_2,
    pub sk_policy: [*mut xfrm_policy; 2],
    pub sk_rx_dst: *mut dst_entry,
    pub sk_dst_cache: *mut dst_entry,
    pub sk_omem_alloc: atomic_t,
    pub sk_sndbuf: c_int,
    pub sk_wmem_queued: c_int,
    pub sk_wmem_alloc: refcount_t,
    pub sk_tsq_flags: c_ulong,
    pub __bindgen_anon_2: sock__bindgen_ty_3,
    pub sk_write_queue: sk_buff_head,
    pub sk_peek_off: __s32,
    pub sk_write_pending: c_int,
    pub sk_dst_pending_confirm: __u32,
    pub sk_pacing_status: u32,
    pub sk_sndtimeo: c_long,
    pub sk_timer: timer_list,
    pub sk_priority: __u32,
    pub sk_mark: __u32,
    pub sk_pacing_rate: u32,
    pub sk_max_pacing_rate: u32,
    pub sk_frag: page_frag,
    pub sk_route_caps: netdev_features_t,
    pub sk_route_nocaps: netdev_features_t,
    pub sk_gso_type: c_int,
    pub sk_gso_max_size: c_uint,
    pub sk_allocation: gfp_t,
    pub sk_txhash: __u32,
    pub __sk_flags_offset: __IncompleteArrayField<c_uint>,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4], u16>,
    pub sk_gso_max_segs: u16,
    pub sk_pacing_shift: u8,
    pub sk_lingertime: c_ulong,
    pub sk_prot_creator: *mut proto,
    pub sk_callback_lock: rwlock_t,
    pub sk_err: c_int,
    pub sk_err_soft: c_int,
    pub sk_ack_backlog: u32,
    pub sk_max_ack_backlog: u32,
    pub sk_uid: kuid_t,
    pub sk_peer_pid: *mut pid,
    pub sk_peer_cred: *const cred,
    pub sk_rcvtimeo: c_long,
    pub sk_stamp: ktime_t,
    pub sk_tsflags: u16,
    pub sk_shutdown: u8,
    pub sk_tskey: u32,
    pub sk_zckey: atomic_t,
    pub sk_socket: *mut socket,
    pub sk_user_data: *mut c_void,
    pub sk_security: *mut c_void,
    pub sk_cgrp_data: sock_cgroup_data,
    pub sk_memcg: *mut mem_cgroup,
    pub sk_state_change: Option<unsafe extern "C" fn(sk: *mut sock)>,
    pub sk_data_ready: Option<unsafe extern "C" fn(sk: *mut sock)>,
    pub sk_write_space: Option<unsafe extern "C" fn(sk: *mut sock)>,
    pub sk_error_report: Option<unsafe extern "C" fn(sk: *mut sock)>,
    pub sk_backlog_rcv: Option<unsafe extern "C" fn(sk: *mut sock, skb: *mut sk_buff) -> c_int>,
    pub sk_destruct: Option<unsafe extern "C" fn(sk: *mut sock)>,
    pub sk_reuseport_cb: *mut sock_reuseport,
    pub sk_rcu: callback_head,
}

struct sock - network layer representation of sockets @__sk_common: shared layout with inet_timewait_sock @sk_shutdown: mask of %SEND_SHUTDOWN and/or %RCV_SHUTDOWN @sk_userlocks: %SO_SNDBUF and %SO_RCVBUF settings @sk_lock: synchronizer @sk_kern_sock: True if sock is using kernel lock classes @sk_rcvbuf: size of receive buffer in bytes @sk_wq: sock wait queue and async head @sk_rx_dst: receive input route used by early demux @sk_dst_cache: destination cache @sk_dst_pending_confirm: need to confirm neighbour @sk_policy: flow policy @sk_receive_queue: incoming packets @sk_wmem_alloc: transmit queue bytes committed @sk_tsq_flags: TCP Small Queues flags @sk_write_queue: Packet sending queue @sk_omem_alloc: "o" is "option" or "other" @sk_wmem_queued: persistent queue size @sk_forward_alloc: space allocated forward @sk_napi_id: id of the last napi context to receive data for sk @sk_ll_usec: usecs to busypoll when there is no data @sk_allocation: allocation mode @sk_pacing_rate: Pacing rate (if supported by transport/packet scheduler) @sk_pacing_status: Pacing status (requested, handled by sch_fq) @sk_max_pacing_rate: Maximum pacing rate (%SO_MAX_PACING_RATE) @sk_sndbuf: size of send buffer in bytes @__sk_flags_offset: empty field used to determine location of bitfield @sk_padding: unused element for alignment @sk_no_check_tx: %SO_NO_CHECK setting, set checksum in TX packets @sk_no_check_rx: allow zero checksum in RX packets @sk_route_caps: route capabilities (e.g. %NETIF_F_TSO) @sk_route_nocaps: forbidden route capabilities (e.g NETIF_F_GSO_MASK) @sk_gso_type: GSO type (e.g. %SKB_GSO_TCPV4) @sk_gso_max_size: Maximum GSO segment size to build @sk_gso_max_segs: Maximum number of GSO segments @sk_pacing_shift: scaling factor for TCP Small Queues @sk_lingertime: %SO_LINGER l_linger setting @sk_backlog: always used with the per-socket spinlock held @sk_callback_lock: used with the callbacks in the end of this struct @sk_error_queue: rarely used @sk_prot_creator: sk_prot of original sock creator (see ipv6_setsockopt, IPV6_ADDRFORM for instance) @sk_err: last error @sk_err_soft: errors that don't cause failure but are the cause of a persistent failure not just 'timed out' @sk_drops: raw/udp drops counter @sk_ack_backlog: current listen backlog @sk_max_ack_backlog: listen backlog set in listen() @sk_uid: user id of owner @sk_priority: %SO_PRIORITY setting @sk_type: socket type (%SOCK_STREAM, etc) @sk_protocol: which protocol this socket belongs in this network family @sk_peer_pid: &struct pid for this socket's peer @sk_peer_cred: %SO_PEERCRED setting @sk_rcvlowat: %SO_RCVLOWAT setting @sk_rcvtimeo: %SO_RCVTIMEO setting @sk_sndtimeo: %SO_SNDTIMEO setting @sk_txhash: computed flow hash for use on transmit @sk_filter: socket filtering instructions @sk_timer: sock cleanup timer @sk_stamp: time stamp of last packet received @sk_tsflags: SO_TIMESTAMPING socket options @sk_tskey: counter to disambiguate concurrent tstamp requests @sk_zckey: counter to order MSG_ZEROCOPY notifications @sk_socket: Identd and reporting IO signals @sk_user_data: RPC layer private data @sk_frag: cached page frag @sk_peek_off: current peek_offset value @sk_send_head: front of stuff to transmit @sk_security: used by security modules @sk_mark: generic packet mark @sk_cgrp_data: cgroup data for this cgroup @sk_memcg: this socket's memory cgroup association @sk_write_pending: a write to stream socket waits to start @sk_state_change: callback to indicate change in the state of the sock @sk_data_ready: callback to indicate there is data to be processed @sk_write_space: callback to indicate there is bf sending space available @sk_error_report: callback to indicate errors (e.g. %MSG_ERRQUEUE) @sk_backlog_rcv: callback to process the backlog @sk_destruct: called at sock freeing time, i.e. when all refcnt == 0 @sk_reuseport_cb: reuseport group container @sk_rcu: used during RCU grace period

Fields
__sk_common: sock_commonsk_lock: socket_lock_tsk_drops: atomic_tsk_rcvlowat: c_intsk_error_queue: sk_buff_headsk_receive_queue: sk_buff_headsk_backlog: sock__bindgen_ty_1sk_forward_alloc: c_intsk_ll_usec: c_uintsk_napi_id: c_uintsk_rcvbuf: c_intsk_filter: *mut sk_filter__bindgen_anon_1: sock__bindgen_ty_2sk_policy: [*mut xfrm_policy; 2]sk_rx_dst: *mut dst_entrysk_dst_cache: *mut dst_entrysk_omem_alloc: atomic_tsk_sndbuf: c_intsk_wmem_queued: c_intsk_wmem_alloc: refcount_tsk_tsq_flags: c_ulong__bindgen_anon_2: sock__bindgen_ty_3sk_write_queue: sk_buff_headsk_peek_off: __s32sk_write_pending: c_intsk_dst_pending_confirm: __u32sk_pacing_status: u32sk_sndtimeo: c_longsk_timer: timer_listsk_priority: __u32sk_mark: __u32sk_pacing_rate: u32sk_max_pacing_rate: u32sk_frag: page_fragsk_route_caps: netdev_features_tsk_route_nocaps: netdev_features_tsk_gso_type: c_intsk_gso_max_size: c_uintsk_allocation: gfp_tsk_txhash: __u32__sk_flags_offset: __IncompleteArrayField<c_uint>_bitfield_1: __BindgenBitfieldUnit<[u8; 4], u16>sk_gso_max_segs: u16sk_pacing_shift: u8sk_lingertime: c_ulongsk_prot_creator: *mut protosk_callback_lock: rwlock_tsk_err: c_intsk_err_soft: c_intsk_ack_backlog: u32sk_max_ack_backlog: u32sk_uid: kuid_tsk_peer_pid: *mut pidsk_peer_cred: *const credsk_rcvtimeo: c_longsk_stamp: ktime_tsk_tsflags: u16sk_shutdown: u8sk_tskey: u32sk_zckey: atomic_tsk_socket: *mut socketsk_user_data: *mut c_voidsk_security: *mut c_voidsk_cgrp_data: sock_cgroup_datask_memcg: *mut mem_cgroupsk_state_change: Option<unsafe extern "C" fn(sk: *mut sock)>sk_data_ready: Option<unsafe extern "C" fn(sk: *mut sock)>sk_write_space: Option<unsafe extern "C" fn(sk: *mut sock)>sk_error_report: Option<unsafe extern "C" fn(sk: *mut sock)>sk_backlog_rcv: Option<unsafe extern "C" fn(sk: *mut sock, skb: *mut sk_buff) -> c_int>sk_destruct: Option<unsafe extern "C" fn(sk: *mut sock)>sk_reuseport_cb: *mut sock_reuseportsk_rcu: callback_head
Methods

impl sock

pub fn sk_padding(&self) -> c_uint

pub fn set_sk_padding(&mut self, val: c_uint)

pub fn sk_kern_sock(&self) -> c_uint

pub fn set_sk_kern_sock(&mut self, val: c_uint)

pub fn sk_no_check_tx(&self) -> c_uint

pub fn set_sk_no_check_tx(&mut self, val: c_uint)

pub fn sk_no_check_rx(&self) -> c_uint

pub fn set_sk_no_check_rx(&mut self, val: c_uint)

pub fn sk_userlocks(&self) -> c_uint

pub fn set_sk_userlocks(&mut self, val: c_uint)

pub fn sk_protocol(&self) -> c_uint

pub fn set_sk_protocol(&mut self, val: c_uint)

pub fn sk_type(&self) -> c_uint

pub fn set_sk_type(&mut self, val: c_uint)

pub fn new_bitfield_1(
sk_padding: c_uint,
sk_kern_sock: c_uint,
sk_no_check_tx: c_uint,
sk_no_check_rx: c_uint,
sk_userlocks: c_uint,
sk_protocol: c_uint,
sk_type: c_uint
) -> __BindgenBitfieldUnit<[u8; 4], u16>